requirements.txt
openai>=1.40.0
google-generativeai>=0.7.2
